Fixed resolution bug, added enemy attack warning and an early version of attack handler for when different enemy types are added

+++ src/GUI_Button.h
@@ -31,7 +31,7 @@ public:
 
     sf::FloatRect getBounds() const;
 
-    bool isClicked(sf::Vector2i mousePos, Key_Manager & KeyManager);
+    bool isClicked(sf::Vector2f mousePos, Key_Manager & KeyManager);
 };
